feat: update swiper
